See desc.
- Added basic file io to read shader files.
- Added SetViewport to RendererAPI.
- Made all window and framebuffer sizes signed.

+++ GBC/src/GBC/Event/KeyEvents.cpp
@@ -7,28 +7,28 @@ namespace gbc
 	std::string KeyPressEvent::ToString() const
 	{
 		std::stringstream stream;
-		stream << "KeyPressEvent: keycode=" << static_cast<uint16_t>(m_Keycode) << " modifiers=" << +m_Modifiers;
+		stream << "KeyPressEvent: keycode=" << m_Keycode << " modifiers=" << +m_Modifiers;
 		return stream.str();
 	}
 
 	std::string KeyRepeatEvent::ToString() const
 	{
 		std::stringstream stream;
-		stream << "KeyRepeatEvent: keycode=" << static_cast<uint16_t>(m_Keycode) << " modifiers=" << +m_Modifiers;
+		stream << "KeyRepeatEvent: keycode=" << m_Keycode << " modifiers=" << +m_Modifiers;
 		return stream.str();
 	}
 
 	std::string KeyReleaseEvent::ToString() const
 	{
 		std::stringstream stream;
-		stream << "KeyReleaseEvent: keycode=" << static_cast<uint16_t>(m_Keycode) << " modifiers=" << +m_Modifiers;
+		stream << "KeyReleaseEvent: keycode=" << m_Keycode << " modifiers=" << +m_Modifiers;
 		return stream.str();
 	}
 
 	std::string KeyCharTypeEvent::ToString() const
 	{
 		std::stringstream stream;
-		stream << "KeyCharTypeEvent: codepoint=" << static_cast<uint32_t>(m_Codepoint);
+		stream << "KeyCharTypeEvent: codepoint=" << m_Codepoint;
 		return stream.str();
 	}
 }
